Split entity packages so alerting can just scan auth entity
