+++ aw-reporting-model/src/main/java/com/google/api/ads/adwords/awreporting/model/persistence/mongodb/MongoDbAuthTokenPersister.java
@@ -14,7 +14,7 @@
 
 package com.google.api.ads.adwords.awreporting.model.persistence.mongodb;
 
-import com.google.api.ads.adwords.awreporting.model.entities.AuthMcc;
+import com.google.api.ads.adwords.awreporting.model.entities.auth.AuthMcc;
 import com.google.api.ads.adwords.awreporting.model.persistence.AuthTokenPersister;
 import com.google.api.ads.adwords.awreporting.model.persistence.EntityPersister;
 
@@ -49,7 +49,7 @@ public class MongoDbAuthTokenPersister implements AuthTokenPersister {
   /**
    * @see com.google.api.ads.adwords.awreporting.model.persistence.AuthTokenPersister#
    *      persistAuthToken(
-   *      com.google.api.ads.adwords.awreporting.model.entities.AuthMcc)
+   *      com.google.api.ads.adwords.awreporting.model.entities.auth.AuthMcc)
    */
   @Override
   public void persistAuthToken(AuthMcc authToken) {
