+++ aw-reporting-model/src/main/java/com/google/api/ads/adwords/awreporting/model/persistence/sql/SqlReportEntitiesPersister.java
@@ -14,9 +14,9 @@
 
 package com.google.api.ads.adwords.awreporting.model.persistence.sql;
 
-import com.google.api.ads.adwords.awreporting.model.entities.AuthMcc;
 import com.google.api.ads.adwords.awreporting.model.entities.Report;
 import com.google.api.ads.adwords.awreporting.model.entities.ReportBase;
+import com.google.api.ads.adwords.awreporting.model.entities.auth.AuthMcc;
 import com.google.api.ads.adwords.awreporting.model.persistence.EntityPersister;
 import com.google.common.collect.Maps;
