+++ aw-reporting-model/src/main/resources/aw-report-sql-beans.xml
@@ -2,12 +2,12 @@
 <beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xmlns:context="http://www.springframework.org/schema/context" xmlns:tx="http://www.springframework.org/schema/tx"
   xsi:schemaLocation="
-	http://www.springframework.org/schema/beans 
-	http://www.springframework.org/schema/beans/spring-beans.xsd
-	http://www.springframework.org/schema/context
-  http://www.springframework.org/schema/context/spring-context.xsd
-  http://www.springframework.org/schema/tx 
-  http://www.springframework.org/schema/tx/spring-tx.xsd
+    http://www.springframework.org/schema/beans 
+    http://www.springframework.org/schema/beans/spring-beans.xsd
+    http://www.springframework.org/schema/context
+    http://www.springframework.org/schema/context/spring-context.xsd
+    http://www.springframework.org/schema/tx 
+    http://www.springframework.org/schema/tx/spring-tx.xsd
   ">
 
   <tx:annotation-driven />
@@ -29,8 +29,12 @@
         <prop key="hibernate.order_inserts">true</prop>
       </props>
     </property>
-    <property name="packagesToScan"
-      value="com.google.api.ads.adwords.awreporting.model.entities" />
+    <property name="packagesToScan">
+      <list>
+        <value>com.google.api.ads.adwords.awreporting.model.entities.auth</value>
+        <value>com.google.api.ads.adwords.awreporting.model.entities</value>
+      </list>
+    </property>
   </bean>
 
   <bean id="transactionManager" class="org.springframework.orm.hibernate4.HibernateTransactionManager">
